move get_vf_dict to DKM
